Move base class to base module
